Added a lot of missing delete operations added: delete operations on all aspects of the package builder added: show/update methods for the options moved: starting to move buttons on the modals where primary action is right, delete is left

+++ app/Models/InvoiceItem.php
@@ -86,7 +86,7 @@ class InvoiceItem extends Model
     {
         $catalogPrice = $this->getCatalogPrice();
         $diff = $this->price / $catalogPrice;
-        return 100 - round($diff * 100);
+        return (int) (100 - round($diff * 100));
     }
 
     /**
