+++ app/Models/QuoteItem.php
@@ -265,7 +265,7 @@ class QuoteItem extends Model
     {
         $catalogPrice = $this->getCatalogPrice();
         $diff = $this->price / $catalogPrice;
-        return 100 - round($diff * 100);
+        return (int) (100 - round($diff * 100));
     }
 
     /**
